refactor: move workflow dispatch into provider
